requirements.txt
python-dotenv 
supabase
requests